pyproject.toml
[project]
name = "questionnaire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.124.4",
    "google-auth>=2.43.0",
    "google-genai>=1.55.0",
    "gspread>=6.2.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pysqlite3>=0.5.4",
    "uvicorn[standard]>=0.38.0",
]
